TINKERPOP-2518 Enable more Gherkin scenarios
There were three categories of scenarios ignored that are now enabled:
* Some scenarios passed without any change. I don't know why they were
ignored in the first place. Maybe we've fixed something in the meantime
but forgot to enable them again.
* Some scenarios expected a decimal type as the result but the server
sent a double. I simply changed the expected type there to reflect what
we get back from the server.
* One scenario required a more complex assertion as it expects an array
of dictionaries which means that we have to perform a deep equality
check.